Fixes thread blocking issue and improves performance.

+++ service/src/main/groovy/org/olf/dataimport/erm/CoverageStatement.groovy
@@ -20,7 +20,7 @@ class CoverageStatement extends AbstractCoverageStatement implements CoverageSta
   String endVolume
   String endIssue
   
-  int version = 0
+  Long version = 0
   
   static constraints = {
     startDate(nullable:false, validator: STATEMENT_START_VALIDATOR)
